requirements.txt
bcrypt==4.3.0
boto3==1.35.19
botocore==1.35.19
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==45.0.5
et_xmlfile==2.0.0
greenlet==3.1.0
idna==3.10
jmespath==1.0.1
numpy==2.0.2
openpyxl==3.1.5
pandas==2.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodomex==3.23.0
PyJWT==2.10.1
PyMySQL==1.1.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
